requirements.txt
flask
pandas
sqlalchemy
psycopg2-binary
redis
python-dotenv
